requirements.txt
requests
pandas
binance-futures-connector
python-dotenv
websocket-client